xdSettings: MediaOutputPanel: Handle the case of missing packageName
If packageName is null here it causes an exception at
MediaOutputDialogReceiver. In case of it being null it
should be safe to assume the currently active media
session.

If there is no currently active media session set packageName
to "" instead of null to avoid the previous exception and
launch the media output dialog without details about the
media session.

Change-Id: Ie0e1a3354450c71c511ae7cdba1ce4747d33174e